feat: migrate Help and ReleaseNotes components to new directory structure with updated content handling
